+++ src/components/sdk/DeviceSelection/MicSelection/index.tsx
@@ -5,6 +5,7 @@ import React from 'react';
 
 import { useSelectAudioInputDevice } from '../../../../hooks/sdk/useSelectAudioInputDevice';
 import { useAudioInputs } from '../../../../providers/DevicesProvider';
+import { DeviceConfig } from '../../../../types';
 import DeviceInput from '../DeviceInput';
 
 interface Props {
@@ -12,14 +13,20 @@ interface Props {
   notFoundMsg?: string;
   /** The label that will be shown for microphone selection, it defaults to `Microphone source`. */
   label?: string;
+  /** A boolean that determines whether or not to include additional sample audio input devices, such as "None", "440 Hz". Defaults to true. This will be deprecated in the next major version. */
+  appendSampleDevices?: boolean;
 }
 
 export const MicSelection: React.FC<Props> = ({
   notFoundMsg = 'No microphone devices found',
   label = 'Microphone source',
+  appendSampleDevices = true,
 }) => {
+  const audioInputConfig: DeviceConfig = {
+    additionalDevices: appendSampleDevices,
+  };
   const selectAudioInput = useSelectAudioInputDevice();
-  const { devices, selectedDevice } = useAudioInputs();
+  const { devices, selectedDevice } = useAudioInputs(audioInputConfig);
 
   return (
     <DeviceInput